Uniffi v0.28.3

+++ bindgen/src/bindings/cpp/templates/str_conv.hpp
@@ -3,5 +3,5 @@ struct {{ typ|ffi_converter_name }} {
     static RustBuffer lower(const std::string &);
     static std::string read(RustStream &);
     static void write(RustStream &, const std::string &);
-    static int32_t allocation_size(const std::string &);
+    static uint64_t allocation_size(const std::string &);
 };
